Refactored null attribute handling to use `CROSS APPLY`
There isn't any performance difference to using `CROSS APPLY`, but I find the syntax a bit cleaner than a nested query.

It's worth noting that I tried caching the existing values in a local `AttributeValues` type as a performance improvement, but that was dramatically slower due to the cost of querying `AttributeIndex`, compared to the very fast subquery against `Attributes`.