Prevent duplicates for indexed attributes
Currently, we rely on the caller to eliminate duplicate values. This works fine. But it'd be nice to prevent that at the database level just to be safe.

This adds a _small_ amount of overhead, but it's relatively trivial compared to the benefit. With this `OUTER APPLY` added, attribute values will _only_ be added if they are different from the last value saved for that key. This ensures we don't introduce unnecessary consecutive duplicates to the attributes table.

This is handy for casual scripting, as we can call `UpdateTopic` with whatever values we want, and ensure they are added if appropriate. If they are already the current value, they will be skipped. This also adds utility for cases where we are importing data, but don't know the current state of the database, as it prevents the need to first query the database, then deduplicate the records, and finally the delta.