Remove `@DeleteRelationships` support in `@UpdateTopic`
Previously, in order to update existing relationships, callers had to trigger `UpdateTopic`'s `@DeleteRelationships` flag—which would delete all relationships for that topic—and _then_ call `UpdateRelationships` with the final relationships set. This was wasteful, as often times the relationships will have overlap, or even be identical.

This issue was fixed with with the introduction of the `@DeleteUnmatched` parameter to the `UpdateRelationships` stored procedure in the last commit (46ba436). Given that, we can now remove the `@DeleteRelationships` functionality from `UpdateTopic`. That said, since `@DeleteRelationships` is part of the public interface, we don't want to remove it as that would require that the `SqlTopicRepository` version be aligned with the database version—something we restrict to major releases, as part of semantic versioning.

As such, we're keeping the `@DeleteRelationships` parameter, but removing the functionality. If the database is up-to-date with this release, that will yield the same results as before. If the database isn't, it will still accept the `@DeleteRelationships` parameter, as expected, and behave as it used to.

(Since the only reasonable purpose of calling `@DeleteRelationships` was to subsequently call `UpdateRelationships`, there's no expected off-target impact to this.)