Renamed `Topic.IsSaved` to `Topic.IsNew`
In a recent commit, we introduced the new `Topic.IsSaved` property to determine if a topic mapped to a topic saved in the underlying `ITopicRepository` (1678523). This is prone to confusion because `IsSaved` could imply that the topic is an existing entity, but that its current state hasn't yet been saved. To help avoid that confusion, I've renamed it to `Topic.IsNew`. This is less ambiguous.

As this has the opposite meaning of the previous property, this means flipping the `!` references.