Refactored logic into `GetNavigationRoot()`, `MapNavigationViewModels()`
The `MenuViewComponentBase<T>` and `PageLevelNavigationViewComponentBase<T>` follow a similar structure, where they first retrieve the navigation root, then use it to map navigation view models.  This logic is now refactored into two `protected` methods, `GetNavigationRoot()` and `MapNavigationViewModels()`.

This not only formalizes a common structure between these, but more importantly allows these to be overwritten by derived view components in case they need to override the logic—or, more likely, alter the data once the base logic has been executed. This is especially important since, due to an apparent bug in how view components are written, the `InvokeAsync()` can't be overwritten (if it is, the Razor templates won't be able to find it).