Introduced the `ChildNavigationViewComponent`
The `ChildNavigationViewComponent` (`<vc:child-navigation />`) exposes all child topics for the current topic. This might be useful for providing navigation at e.g. the bottom of the page in order to drill down to lower-level pages. It might also be useful for index pages, in order to display a list of all pages under the current index.