Updated documentation for `MapTopicRoute(IEndpointRouteBuilder)`
Updated the documentation to include the actual pattern being implemented, so it's less obscured by the extension method. This is a best practice that Microsoft implements with many of their own extension methods, and which we should adopt.

In addition, removed the reference back to the `MapTopicRoute(IRouteBuilder)` extension method, since it's now marked as deprecated.