Introduced `MapTopicAreaRoute()` extension method.
This is identical to the existing `MapTopicRoute()` extension method, but it accounts for routes. It accepts an `areaName`, and will default to a `controller` of the same name if not otherwise provided. (In this way, `areaName` is analogous to `rootTopic`—and, indeed, will be set to `rootTopic` in the `defaults`.) An `action` can also be supplied, otherwise it will default to `Index`.