Introduced `MapImplicitAreaControllerRoute()`
The `MapImplicitAreaControllerRoute()` extension method is similar to the `MapDefaultAreaControllerRoute()` recently introduced (35cdd74), except that it supports a short-hand convention for scenarios where the controller name is the same as the area name. This allows e.g., a `[Area("Forms")] public FormsController: Controller` to be routed via `/Forms`, instead of the redundant `/Forms/Forms`.

This convention is most commonly used for scenarios where there is a single controller for an area.