Introduced optional `includeMetadata` parameter
The `includeMetadata` parameter—which defaults to `false`—allows the inclusion of extended metadata such as (a limited set of) topic attributes as well as topic relationships to be included in the sitemap. These were previously supported as they could be used by Google Custom Search Engine (CSE) in order to provide site-specific queries. While that was a really useful feature, CSE is no longer supported by Google, and thus this functionality is no longer used.

(Technically, this is a breaking change. As no APIs that implementers are using currently support or rely on this, however, we're not worried about that change.)

As other agents or search engines may, in the future, offer support for extended attributes via the `<PageMap />` and `<DataObject />` elements, I've kept that functionality in via the optional `includeMetadata` parameter, thus allowing this to still be used if needed.

The immediate affect of this is to (dramatically) reduce the size of the default `/Sitemap` route, since it wil no longer include metadata for attributes and relationships on every topic.