Introduced new `Extended` action
The new `Extended()` action is simply a shortcut to the previously introduced `includeMetadata` (87b6d1b), which will, by default, be routed from `/Sitemap/Extended` (assuming e.g. `MapDefaultControllerRoute()`).

This will output the same sitemap as was previously output, with `<PageMap>` and `<DataObject>`s representing (most) attributes and relationships for every topic. This dramatically increases the size of the XML, but provides useful metadata for any agents that might support it.