Introduced new `MapTopicSitemap()` extension method
Normally, the sitemap will be routed using the standard route—e.g., as supported by `MapDefaultControllerRoute()`. If that isn't being used, for some reason, the `MapTopicSitemap()` method provides a convenience method for establishing a specific route for the sitemap.

Currently, this hardcodes the route to `/Sitemap`. In the future, we may choose to extend that to provide a variable path, but that's not an immediate priority.