Introduce `ContentTypeDescriptorCollection.Refresh()` method
The `Refresh()` method makes it easier for the `ContentTypeDescriptionCollection` to be used as a cache for the flatterned version of the `ContentTypeDescriptor`s used for the `ITopicRepository.GetContentTypes()` cache—which is it's primary use case today. Specifically, the `Refresh()` method allows callers to send a `rootContentType` (i.e., `Root:Configuration:ContentTypes`) to the `Refresh()` method in order to update the cached values.