Updated unit test to pass root content type descriptor
Previously, the `GetContentTypeDescriptors_WithTopicGraph_ReturnsMergedContentTypes()` unit test validted its functionality by passing in a specific content type. This worked, but only because the initial functionality failed to validate that the `ContentTypeDescriptor` being sent was, in fact, the expected root `ContentTypeDescriptor`. This issue will ultimately be fixed by using the new `ContentTypeDescriptorCollection.Refresh()` method (3e3930f)'s validation (b206f3e). In anticipation of migrating to that, the unit test is being updated to be more specific.