Introduced new `SetContentTypeDescriptors()` method
Previously, the `GetContentTypeDescriptors(ContentTypeDescriptor)` method provided a way of retrieving the content type descriptor cache, but also a way of updating it. In practice, in most of the places it was used, it was semantically being used purely to update the cache, either because a content type was missing, or because the metadata was being modified.

Given that, it makes more sense to rename it to `SetContentTypeDescriptors()`. Because `GetContentTypeDescriptors()` was already published as part of the public interface, however, it's being marked as deprecated via `[Obsolete()]`, and will be removed in the next major version of the library (5.0.0).