Introduced new `SetContentTypeDescriptors()` overload
The initial overload of `SetContentTypeDescriptors()` (f1ade10) operated just like the `GetContentTypeDescriptors()` method it replaced in that it accepted a `ContentTypeDescriptor`—which it expected to be the _root_ `ContentTypeDescriptor` (i.e., `Root:Configuration:ContentTypes`) (b206f3e)

This overload provides a bit more flexibility by allowing a caller to submit _any_ `Topic`—not just a `ContentTypeDescriptor`—from _anywhere_ within the topic graph—not just `Root:Configuration:ContentTypes`. This is just a convenience method to make it really easy to update the content type configuration by passing in a reference to the topic graph via any node in that graph, and trusting it to find the node it relies on.

(This process is easy with the newish `GetByUniqueKey()` extension method, but it's still handy having it centralized, instead of needing to do that lookup each time it's called.)