Removed redundant `ContentTypeDescriptor` update in `Save()`
The `Save()` method currently has two places where it updates the `_ContentTypeDescriptors` cache. This is redundant, and means the second one—which is actually more cautious—will never actually fire. To avoid this, I've removed the first one.

While I was at it, I also updated the second one to use pattern matching syntax for the variable definition, which is a bit less code, and easier to read.