Fixed bug in `Move()` related to updating `AttributeDescriptors`
When moving a `ContentTypeDescriptor` to a different parent, the `ContentTypeDescriptor.AttributeDescriptors` collections of each descendant `ContentTypeDescriptor` must be updated, since those will contain cached references to `AttributeDescriptor`s inherited from their parents.

While this code was in place, it compared the `topic.Parent` to the `target` (parent) _after_ `topic` had _already_ been moved to `target`. As such, this code would never fire, because it would never appear as though the `ContentTypeDescriptor` was being moved to a _new_ parent. Oops.

This fixes this by storing a reference to the previous parent, and then comparing that to the target after calling `SetParent()`.