Renamed `GetContentTypeDescriptorsProxy()`
Since `TopicRepositoryBase.GetContentTypeDescriptors(ContentTypeDescriptor)` has been marked as deprecated, in preference for the new `SetContentTypeDescriptors()` method (f1ade10), the `GetContentTypeDescriptorsProxy()` needs to be updated to call the new method.

While I was at it, I also deprecated the previous proxy itself, and introduced a new proxy that maps to the new name. This is perhaps not necessary since this is exclusively used by the `OnTopic.Tests`. The `StubTopicRepository` is publicly published via the `Ontopic.TestDoubles` package, however, so that it can be shared with other OnTopic related projects, such as `OnTopic-MVC`. This will help ensure compatibility with those projects.