Initialize `GetContentTypeDescriptors()` cache on `Load()`
When `SqlTopicRepository.Load()` is called, attempt to update the `GetContentTypeDescriptors()` cache by calling the new `SetContentTypeDescriptors()` overload (78a9d3f).

There are two main advantages to this. First, it helps prevent an unnecessary and redundant call to the database in order to initialize the `GetContentTypeDescriptors()` cache. Second, it helps further ensure that the object references used by that cache are the same ones used by any subsequent caching layer since, in practice, the first call to `Load()` will be to warm the site cache by pulling the entire topic graph.

(Unfortunately, because `CachedTopicRepository` uses a decorator pattern, it can't otherwise share the cache very easily with the underlying `ITopicRepository` instance, since that instance isn't guaranteed to inherit from `TopicRepositoryBase`—and since it isn't using actual inheritance. The decorator pattern allows the `CachedTopicRepository` to be used with _any_ implementation, not just the e.g., `SqlTopicRepository`, so this makes sense. But given that, we need to get creative to avoid versioning conflicts between caches. In most cases, this update should mitigate that.)