Ensured `Move()` overload properly functions
This is embarrassing, but it seems the basic `Move()` overload in `TopicRepositoryBase` doesn't refer to the more complete `Move()` overload, nor does it do so in any of the concrete implementations. As a result, it is effectively unimplemented, and won't persist any changes to the database.

This is fixed here by simply referring it to the more complete overload (which additionally accepts `sibling`).

This also exposes a bigger issue: Most of these `public` methods should be marked as `abstract` since they're not fully implemented, and just intended to provide common support. That breaks the public contract, though, so will need to wait for 5.0.0.