Fixed errant unit test which covered up previous bug
There was a bug in the `Move_ContentTypeDescriptor_UpdatesContentTypeCache` unit test, which is one of the reasons we hadn't noticed the previous bug sooner. That unit test is now broken, appropriately, since the functionality now works. I've fixed the logic in the unit test so it correctly evaluates this functionality.