Mitigated validation bug with recent `Move()` bug fix
When fixing the missing implementation of the basic `Move()` overload (d7a61cd), I failed to take into account that `SetParent()` validates that the sibling is not null. Perhaps we should align the logic of `SetParent()` to behave like `Move()` where `sibling` can be `null`. In the meanwhile, however, it's easy to workaround.