Introduced `ResetPermittedContentTypes()`
Similar to the previously introduced `ResetAttributeDescriptors()` (b0c5a5ec), this exposes a new `internal` helper method, `ResetPermittedContentTypes()`, which is called from `TopicRepositoryBase.Save()` when a `ContentTypeDescriptor` is updated, assuming its relationships have been updated.

In practice, this might be called rarely enough, and may be fast enough to assemble, that we could just do away with the cache entirely. That said, this is a familiar and easy pattern to apply alongside `ResetAttributeDescriptors()`, so we'll stick with this for now.

Also created a corresponding unit test to validate this functionality.