Explicitly refresh content type cache on `Move()`
Despite multiple adustments to how `Move()` handles `ResetAttributeDescriptors()`, this functionality still isn't working in live site integration testing, even though it seems to work in both unit tests and local integration tests. That likely suggests a discrepency in the lifestyle configuration of dependencies, which could be introducing a version conflict.

This _should_ be mitigated by other measures I've put in place which help ensure that any caching layer (e.g., `CachedTopicRepository`) is using the same object references as the `TopicRepositoryBase.GetContentTypes()`.

Since that's not working, I'm going to try _explicitly_ updating the content type cache via `Move()`, just to be safe. This should not be necessary—though it won't really hurt anything, either. (There's a small amount of overhead, but `Move()` operations are rare, and the database cost will far outweigh this extra validation.)