Throw a `403` error if a `Container` is requested
`Container` topics are topics used to help organize topics. They are not meant to be viewed directly, and are not expected to have pages associated with them. In a recent update, we excluded `Container` topics from the `Sitemap` (87b6d1b), so they shouldn't be referenced, but they may still be discoverable as interstitial folders.