Introduced `GetTopicIDByUniqueKey`
The `GetTopicIDByUniqueKey` function accepts a `@UniqueKey` parameter (e.g., `Root:Configuration:ContentTypes`) and looks up the `TopicID` associated with it.

It is similar in concept to `GetTopicID`, except that function naively looks for any topic matching a specific `Key` and returns the first match—which may not be the _expected_ match if the database has multiple topics with the same key.

In a future version, we'll likely replace `GetTopicID` with `GetTopicIDByUniqueKey`. Since that is a breaking change, however, that decision is being deferred to the next major version (i.e., 5.0.0).