Updated `SqlTopicRepository.Save()` to use `UtcNow` for `@Version`
Our servers are all set to `UTC`, so `DateTime.Now` _is_ `DateTime.NowUtc` when working in a staging or live environment. Usually, however, development machines are running in local time—e.g., PDT/PST.

This can cause problems when executing scripts or making changes via the OnTopic Editor from a development environment as versions will be returns as 7-8 hours prior to the server time. Not only is this inconsistent and confusing, but it can yield bugs for scenarios where another version of a topic was saved in the last 7-8 hours.

By using `DateTime.UtcNow` for `@Version`, we mitigate that issue, and help ensure consistency independent of the web server's timezone.