Introduced new `GetContentTypeDescriptor()`
Previously, `ContentType` was a strongly typed reference to an actual `ContentTypeDescriptor` object. This was removed in 4.0.0 in order to facilitate dependency injection, as otherwise `TopicFactory` needed a static dependency upon an `ITopicRepository` instance in order to lookup references.

The new `GetContentTypeDescriptor()` reintroduces this basic capability by attempting to locate the corresponding `ContentTypeDescriptor` within that same topic graph by relying on the existing `GetByUniqueKey()` and `FindFirst()` extensions. This could easily be done without this, but this makes it easier to dynamically look these up.

This potentially allows for code to rely on `AttributeDescriptor` settings in order to customize the output based on configuration settings. Traditionally, these configuration settings were primarily used for the OnTopic Editor for establishing how data was collected and validated. But there's no reason it can't _also_ be used to configure how attribute values are utilized in certain scenarios.