Introduced unit test for `GetContentTypeDescriptor()`
In 026e7a3, I introduced the new `GetContentTypeDescriptor()` extension method. This commit introduces two unit tests to confirm that functionality.

As part of that, it introduces a dependency upon `ITopicRepository` for the `TopicQueryingTest` which wasn't used previously. This isn't strictly necessary, but since the `StubTopicRepository` already has `ContentTypeDescriptor` objects stubbed, this makes it easier to establish a realistic topic graph to test against without duplicating too much scaffolding.