Facilitated extensibility in `PageLevelNavigationViewComponentBase<T>`
While it is expected that `PageLevelNavigationViewComponentBase<T>` would be extended, this was exclusively to set the value of `T`. Derived classes couldn't extend `InvokeAsync()` due to a bug in ASP.NET Core which prevents view components from being called as tag helpers if the `InvokeAsync()` class is defined in a subclass and then overridden in a derived class.

To give derived classes more flexibility, this update introduces flexibility by moving the core business logic out of the `InvokeAsync()` method and into two `protected` methods: `GetNavigationRoot()` and `MapNavigationTopicViewModels()`—the two key actions taken by any navigation view component.

The `InvokeAsync()` still can't be overwritten. But its functionality is now pretty minimal, and its core business logic is exposed for overrides.