Suppress `CA2249` globally
`CA2249` suggests "Consider using 'string.Contains' instead of 'string.IndexOf'". This is good advice. Unfortunately, since we're (currently) targeting .NET Standard 2.0, however, we can't use `Contains()` because it doesn't have an overload supporting `StringComparison`, and is thus not globalization friendly.

(We could implement `Contains()` without `StringComparison`, but that will correctly trigger a different warning.)

We can reevaluate this once we migrate to .NET 5.0 with OnTopic 5.0.0, as then we will no longer be supporting .NET Standard.