Suppress `CS8777`
An update to code analysis throws `CS8777` if a parameter marked `[NotNull]` isn't guaranteed not to be `null` on return. That seems like a good policy. Except in this case, the methods call into `Requires<>()`, which will throw an exception if `requiredObject` is `null`. Roslyn's static code analysis isn't able to detect that, however, so we're suppressing those warnings.