Introduce `AddOrReplace()` method
Sometimes, derived `ITypeLookupService` implementations will want to add a type which has the same name as a previously added type. Notably, this will happen when there is a reference implementation—such as `DefaultTopicLookupService`—but a specific site needs to customize one or more specific types.

In this case, `AddOrReplace()` allows the type to be added if it's missing, and otherwise replaced with the new type reference if it exists.

There aren't any local calls to this. By design, it's something that should only be needed in derived classes implemented by sites using the OnTopic library.