Introduced `TryAdd()`
When adding a type, it is conceivable that a class being derived from has already added that type. For this reason, most derivatives of `StaticTypeLookupService` implement local checks to avoid a key conflict.

Since this is such a common scenario, that logic is being centralized as a `TryAdd()` method. This method will add the `Type` if no `Type` with the same key exists. Otherwise, it will simply return `false`, instead of throwing an exception. If an exception is preferred in this scenario, the existing `Add()` method will satisfy that condition.

This complements the newly added `AddOrReplace()` method, which will replace the existing `Type` if it exists (cd92791).