Introduced new `ReferentialIntegrityException`
The new `ReferentialIntegrityException` derives from `TopicRepositoryException`, and fires when an operation would introduce a referential integrity error in the underlying database.

Currently, this is expected when a topic is deleted which other topics derive from. As such, an optional overload accepts a `Topic` and completes the message based on that `Topic` as well as its `DerivedTopic`.