Validated derived topics on `Delete()`
Previously, a `Delete()` operation would delete a topic and all descendants without any validation of external references. That's fine when it comes to e.g. relationships, as those _should_ be deleted when a target topic is deleted. This raises a concern, however, when an external topic _derives_ from an affected topic, as that could potentially destabilize that topic be depriving it of required attributes.

This update throws a `ReferentialIntegrityException` when this occurs. As this derives from the existing `TopicRepositoryException`, this won't have any impacts on backward compatibility; i.e., the same base exception is being thrown, and can be caught, in the case of a repository exception.

In addition, two unit tests are introduced. The first validates that a topic can still be deleted if the derived topics are descendants. The second validates that the `ReferentialIntegrityException` is thrown if a derived topic is not a descendant.