Validated descendant topics on `Delete()`
Previously, while `Delete()` had a(n optional) parameter named `isRecursive`, it wasn't actually validated; instead, a `Topic` was always deleted, even if it had descendants.

Now, validation is included so that `Delete()` will throw an exception if a topic has descendants _and_ `isRecursive` is set to `false`. That's functionality that should have always been in place, as otherwise the `isRecursive` parameter is misleading and potentially dangerous.

Note that nested topics are considered part of the parent topic, and not evaluated as part of the descendant check. A topic exclusively containing nested topics as children will be deleted.

Includes unit tests to confirm that `isRecursive` is correctly validated when there are descendants, and correctly ignores nested topics.