Set `ITopicRepository.Delete()`'s `isRecursive`
In a previous commit, I introduced validation of `TopicRepositoryBase.Delete(…, isRecursive)`, which had not been previously validated (a0cc21c). This was a major bug in that code expected that a topic wouldn't be deleted if it had descendants would, in fact, still delete the topic and all descendants. Oops.

While this was fixed, it breaks backward compatibility since `isRecursive` was optional. As such, any code calling into `TopicRepositoryBase.Delete()` without explicitly setting `isRecursive` would not have been expecting an exception in these cases.

To "fix" this, I'm setting the `isRecursive` default to `true`. This is technically a breaking change in terms of the interface. But it avoids the bug fix from introducing a breaking change. We'll want to change this back to `false` in OnTopic 5.0.0.

Finally, to maintain forward compatibility with OnTopic 5.0.0, and for clarity, I explicitly set the `isRecursive` flag on all calls to `Delete()` from within the `OnTopic.Tests`.

As we currently maintain a good understanding of the cases where this is implemented, we have high confidence that this won't impact current implementations. Nevertheless, mea culpa!