Consolidated `Move()` overloads
The `Move()` overloads in `TopicRepositoryBase` can easily be consolidated into one method with an optional `sibling` parameter. Not only that, but this is actually how the `ITopicRepository.Move()` is defined, as well as how each of the `ITopicRepository` implementations were implemented. Given this, having the overload only added potential ambiguity since it conflicted with the `Move()` overload in the concrete implementations.

In practice, because the concrete implementations used the consolidated method with an optional `sibling` parameter, this didn't introduce any actual problems, as it simply routed calls to that `override` with `null`. But, theoretically, this could have introduced problems if a concrete implementation had used overloads. It could also have introduced an infinite loop if a derived class accidentally called the overload without `sibling` since that would simply call back into the `override`n method.

As we currently maintain a good understanding of the cases where this is implemented, we have high confidence that this won't impact current implementations. Still, by fixing this, we're avoiding any future confusion or potentially bugs.