Ensured `sibling` is optional in `ITopicRepository.Move()` test doubles
While this isn't strictly required for the current tests, it helps ensure parity with both the interface as well as the other concrete implementations of `ITopicRepository`, and provides the full set

It's worth noting that in absence of this, C# ensures the default is maintained. If the `sibling` parameter is excluded, then the compiler passes the default value of `null` as defined from the `ITopicRepository` definition. As such, optionality becomes implicit in concrete implementations.

Also fixed an XML reference while I was at it.