Remove relationships when calling `TopicRepositoryBase.Delete()`
Previously, when a topic was deleted, relationships from the topic and its descendants were not deleted.

This potentially introduces problems as any target relationships outside of the deleted topic graph would maintain `IncomingRelationships` back to the deleted topics. That means that those topics would persist in memory. It also means that any views displaying content from `IncomingRelationships` may provide links to resources that no longer exist.

To mitigate this, when `Delete()` is called, it now loops through all topics in the topic graph to be deleted, and removes all relationships from them. And, thanks to a new bug fix which ensures reciprocal `IncomingRelationships` are deleted when `RemoveTopic()` is called (f8b80ec), this fully addresses this issue.

In addition, a unit test has been introduced to validate this functionality.