Limited `GetUniqueKey` function to latest `Key`
The original version of `GetUniqueKey()` (b38bb89) would return malformed unique keys in cases where a topic—or any of its descendants—had been renamed. In this case, it would bind not only to the current key, but also the legacy key, thus including both of them in the path. This can easily be fixed by limiting the `CROSS APPLY` to only the latest version.